Use column.comment from catalog/database if column.description is missing
This will display the column description fetched from the database if there is no column description in the DBT config.
Partially fixes #128, partially because the table description is still missing.
Adding the table description would require more substantial changes, since the table description (as far as I could tell) is not found in catalog.json at the moment.